Optimize the input state-vector copy into the LGPU (#1071)
**Context:**
After running different algorithm with LGPU and perform a memory
profile. Show a memory bottleneck for LGPU on the Python layer because
the peak of memory is 3 times the need for the computation.


![image](https://github.com/user-attachments/assets/24eebcf0-49f2-45e1-a63f-71da26cb8dd4)

**Description of the Change:**
Remove tmp allocation and skip indexes computation for common cases.

* Remove temporal GPU allocation for input values and indexes.
* The input state vector is copied directly from the host if **the
target wires are contiguous and start in the most/least significant
wires** (which are the most common cases).
* In the case of custom target wires, LGPU follow the previous algorithm
but with a speedup in the index computation thought parallel computing
with OpenMP


**Benefits:**
Using a test algorithm with 31 qubits produce the following memory
profile:

![newplot(3)](https://github.com/user-attachments/assets/8c74bfb7-f7ed-4759-bd7d-54e23e8e23df)

Reduction of the memory peak from 100GB to 66GB

Note: `memray` measures all the memory allocation, even for the GPU
`cudaMallocX`.

Using the following toy circuit 
```python 
   state_init =  random_normalize_sv(wires-1)
   target_wires = wires[:-1]
   dev = qml.device("lightning.gpu", wires=wires)

    def circuit():
        qml.StatePrep(input_state, wires=target_wires)
                
        return qml.expval(qml.PauliZ(0))
```

Produce the following times 


![image](https://github.com/user-attachments/assets/9fff27a9-bd3b-42b4-9dd6-88ce3119ee07)


![image](https://github.com/user-attachments/assets/e151ed5e-4a0f-445e-af1c-3de502617f44)


**Possible Drawbacks:**



**Related GitHub Issues:**
[sc-58833]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>