fix measurements with dynamic wires (#1081)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently when using dynamically determined wire count (e.g.
`qml.device("lightning.qubit")` or `qml.device("lightning.qubit", shots
= x)`), we have an issue when we perform measurements without specifying
any observables or wires, i.e.

- qml.probs()
- qml.counts()
- qml.sample()

incorrect result is returned, since the number of wires is not known.

**Description of the Change:**
For measurements with no wires (i.e. no observables or wires provided),
we edit the measurement process to include the number of wires. This is
fixed for with and without shots.

**Benefits:**
Get correct results.

**Possible Drawbacks:**

**Related GitHub Issues:**
fixes https://github.com/PennyLaneAI/pennylane-lightning/issues/1086

[sc-85837]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>