Remove unnecessary reset_state in execution when state_vector is freshly initialized (#1076)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Previously, when we perform `execute` of a circuit, we reset the
statevector within `simulation`. However, the statevector is just
freshly initialized when `execute` is called (with the recent changes to
allocate state vector dynamically), so the reset_state is redundant.

**Description of the Change:**
We remove `reset_state` within simulation. The only occasion when
`reset_state` is necessary is when we reuse the statevector (when
multiple circuits to be executed have the same num_wires`) instead of
initializing a new one, hence this is moved to
`dynamic_wires_from_circuit`.

**Benefits:**
No redundant reset state call - improved performance.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-81780]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>