Hide anonymous namespaces in docs (#1097)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

**Description of the Change:**
Hide anonymous namespaces in docs

**Benefits:**
Tidy up the docs

**Possible Drawbacks:**
n/a

**Related GitHub Issues:**
[sc-85484]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>