Lightning implements the `jaxpr_jvp` method (#1087)
**Context:** We implemented a `jaxpr_jvp` method for `lightning.qubit`.
The same method for `default.qubit` was introduced in [this
PR](https://github.com/PennyLaneAI/pennylane/pull/6875). *We will be
using the tape pipeline for the internals for now*. This can be improved
in the future.

**Description of the Change:** As above.

**Benefits:** More `jaxpr` derivatives.

**Possible Drawbacks:** This implementation based on the tape conversion
has several limitations. Compared to the implementation in
`default.qubit` we can support multiple trainable parameters and opmath
operators, but on the other hand, we don't have a way to infer the
trainable parameters in the conversion from the `jaxpr` to a tape. As a
consequence, we assume that every argument is a trainable parameter and
we raise an error if these do not match exactly the trainable parameters
inferred by the converted tape.

**Related GitHub Issues:** None.

[sc-86224]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Christina Lee <chrissie.c.l@gmail.com>
Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>