Fix doxygen generators (#1096)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently Doxygen generates mangled/random IDs for various data which
appears in the C++ API doc - a lot of these are not necessary for users
and can be removed. These also create mismatch with names generated from
other pipelines.

**Description of the Change:**
Remove C++ API docs for:
- variable
- typedef
- struct
- define
- enum


**Benefits:**
Concise C++ API docs
-[links-removed](https://github.com/user-attachments/files/19319029/links-removed.txt)


**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-85483]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>