Use cuquantum API for computing expval(pauli_string) for LGPU (#1104)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently when calling `expval(pauli_string)` in LGPU, we use the
general matrix observable computation. However, the native C++ function
to call `expval(pauli_string)` is already available (which uses
`custatevecComputeExpectationsOnPauliBasis`), but not called from the
Python layer.

**Description of the Change:**
For LGPU, if the observable in an expval is a pauli string, we call the
native C++ function.

**Benefits:**
Significant speedup for expectation value calculation of large
hamiltonians. For example

| Molecule | Qubits | Num terms in hamiltonian | timing of master branch
(s) | timing of updated branch (s) |

|----------|--------|--------------------------|-----------------------------|------------------------------|
| H10 | 20 | 7151 | 2.62 | 1.34 |
| CO2 | 30 | 16170 | 5599.27 | 60.4 |

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-85843]

---------

Co-authored-by: Luis Alfredo Nuñez Meneses <alfredo.nunez@xanadu.ai>
Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>