Add support for Python 3.13 wheels (#1001)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** Add support for building Python 3.13 wheels.

**Description of the Change:** As above.

**Benefits:** Support Python 3.13.

**Possible Drawbacks:** Additional CI overheads. 
Since PennyLane requires Lightning to be installed, we cannot test
PennyLane 3.13 until a Lightning wheel is available.
As PennyLane also pins numpy<2.1, this pin needs to be relaxed as 2.1 is
the first version to support Python 3.13. This will likely require a
"fail-and-merge" process to get the wheels out for Lightning initially.
UPDATE: PennyLane now unpins the Numpy version, so this should be
unblocked.
**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>