Implement PSWAP gate on all PennyLane-Lightning devices (#1088)
**Context:**
Currently, PSWAP does not have a custom implementation within
PennyLane-Lightning. This results in it using the implementation from
PennyLane-Core whenever the gate is used. By implementing the gate
directly in PennyLane-Lightning, performance of the PSWAP gate can be
improved.

**Description of the Change:**
Implements the PSWAP gate natively in Lightning-Qubit, Lightning-Kokkos,
and Lightning-GPU, along with the generators for the PSWAP gate.

**Benefits:**
Native implementation of the PSWAP gate provides significant performance
benefits on all tested PennyLane-Lightning backends. The charts below
show the execution time of circuits each containing 200 PSWAP gates on
varying number of qubits, as well as the realized speedup from the
native implementation.

**NOTE** The Lightning-Qubit backend was run with the dispatcher turned
off to avoid any differences in speedup coming from running AVX kernels
in the original version of the code.

**Benchmarks for a constant number of PSWAPS:**

![pswap-benchmark-constant](https://github.com/user-attachments/assets/2f52767d-3fa2-40a1-be1f-a3ee3d69a0d0)


![pswap-benchmark-speedup-constant](https://github.com/user-attachments/assets/6cacaa97-4dc7-410c-a0bb-1c418a14d134)

**Benchmarks for a dynamic number of PSWAPS, increasing with circuit
size:**

![pswap-benchmark-dynamic](https://github.com/user-attachments/assets/7d7deed6-fcd8-4ed1-a285-7f43d4e6df75)


![pswap-benchmark-speedup-dynamic](https://github.com/user-attachments/assets/666bc7fd-e865-4bcc-8b92-6e857c7dfe37)


**Possible Drawbacks:**

**Related GitHub Issues:**
#976: PSWAP gate implementation in Lightning Qubit
[sc-77445]
[sc-85797]
[sc-85802]
[sc-86524]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>