Fix decomposition of controlled SProd operators (#1120)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
An error when attempting to decompose controlled SProd operator (e.g.
`Controlled(-1 * Z(0), control_wires=[2])`) is reported.

**Description of the Change:**
Add C(SProd) to stopping condition for all devices, so that this is
supported in lightning devices by converting to QubitUnitary matrix.

**Benefits:**
Closes https://github.com/PennyLaneAI/pennylane-lightning/issues/900
Support controlled SProd operator

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-73351]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>