Improve Lightning Kokkos evaluation of expectation value of Pauli strings (#1126)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently when evaluating expval(pauli string) in LK, we copy the
statevector then apply the operation in place, then obtain the
expectation value, which is slow and uses extra memory.

**Description of the Change:**
Compute expectation value directly instead of applying the operation.

**Benefits:**
Time and Memory reduced e.g. when computing the Hamiltonian of the
following molecules:
_CPU:_
| OMP num threads | Molecule | Qubits | Num terms in hamiltonian |
timing of master branch (s) | timing of updated branch (s) |

|-----------------|----------|--------|--------------------------|-----------------------------|------------------------------|
| 1 | C2 | 20 | 2951 | 78.2 | 7.6 |
| 2 | C2 | 20 | 2951 | 79.9 | 4.4 |
| 4 | C2 | 20 | 2951 | 42.0 | 2.2 |


_GPU:_
| Molecule | Qubits | Num terms in hamiltonian | timing of master branch
(s) | timing of updated branch (s) | memory of master branch | memory of
updated branch |

|----------|--------|--------------------------|-----------------------------|------------------------------|-------------------------|--------------------------|
| C2H2 | 24 | 6401 | 22.9 | 3.28 | 1.00GB | 257MB |

**Possible Drawbacks:**

**Related GitHub Issues:**
 [sc-88612]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>