Update Catalyst plugin with breaking `QuantumDevice` interface changes (#1143)
**Context:** We are updating the `QuantumDevice` interface for Catalyst
device plugins, as announced in release
[0.10.0](https://docs.pennylane.ai/projects/catalyst/en/stable/dev/release_notes.html#release-0-10-0).
The PR in Catalyst is currently open here:
https://github.com/PennyLaneAI/catalyst/pull/1680

To perform the update, I would propose merging this PR in lightning
first, and then update the Lightning dependency in Catalyst to run all
our integration tests. I have tested the changes locally with LQ and LK,
but cannot test LGPU on my mac.

**Description of the Change:** 
- removes `PrintState`, `One`, `Zero` methods from the interface
- removes `shots` parameter from `Counts` and `Sample` interface methods
- updates the `LightningSimulator` header to not use macros
- the catalyst PR also removes the shots kwarg from the device
constructor parameters, but I believe lightning was already compliant
with that (the shots must now be obtained from the `SetDeviceShots`
call).

**Benefits:** compatibility with the next release of Catalyst

**Possible Drawbacks:** incompatibility with the previous release of
Catalyst

**Related GitHub Issues:**
[sc-81294]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>