Adding Carryforward flags (#1144)
**Context:**
Pennylane Lighting changed the structure of its codecov test to only run
on test paths with files that changed. This led to
[issues](https://app.codecov.io/gh/PennyLaneAI/pennylane-lightning/pull/1081/blob/pennylane_lightning/lightning_tensor/_measurements.py?dropdown=coverage)
with false positives due to missing test reports. One solution is to use
the [carryforward
flag](https://docs.codecov.com/docs/carryforward-flags) which compares
new reports to previous commits with the flag.

**Description of the Change:**
Adding `carryforward` flag to codecov configuration. Adding `unit_test`
label to all codecov tests

**Benefits:**
Codecov will compare commits reports to previous base commit with a
carryforward flag on cases where new reports are not generated due to no
changes in the relevant files.

**Possible Drawbacks:**
carryforward works best when using codecov flags, there are no flags in
this repo so the carryforward is applied to everything.

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>