Fix x86 macos wheel build (#1163)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Since adding jax==0.5.3 to requirements-tests.txt, macos wheel builds
for x86 has been
[failing](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/15030269152/job/42242462590).
This is because after building the wheel we install
requirements-tests.txt to test the wheel, and jaxlib 0.5.3 no longer has
wheels for x86 MacOS.

**Description of the Change:**
Update requirements-tests.txt to only install jax==0.5.3 for mac that
are not x86. requirements-dev.txt is updated as well.

**Benefits:**
[Wheels now build
again](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/15053788416/job/42314776574?pr=1163).

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-91413]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>