Fix the slow down of the CI caused by the qml.concurrency update in Pennylane (#1174)
# Addressing CI Timeouts 

This PR aims to tackle the recent and recurring CI job timeouts we've
been seeing on the `latest/latest` branch. We've noticed a significant
and unexplained slowdown in our continuous integration pipelines,
causing many actions to fail.

Here are a few examples of the affected workflows:

*
https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/15381144467/job/43272216165
*
https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/15381144467/job/43272279979
*
https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/15381144467/job/43306798534

These are happening since the merge of the following PR in Pennylane:
* https://github.com/PennyLaneAI/pennylane/pull/7401

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>