Refactor MPIManager from GPU (#1162)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Currently the MPIManager class uses lots of LGPU-specific data/methods,
which prevents reusing for other devices. This PR separates out a base
MPIManger class, which does not contain any cuda specific information,
and a derived MPIManagerGPU class which adds back the necessary cuda
specific information for LGPU-MPI.


**Description of the Change:**
- Remove references to cuda/GPU in MPIManager class
- Create new MPIManagerGPU class which includes cuda/GPU information
- Move MPIManager file to utils, and keep MPIManagerGPU in cuda_utils

**Benefits:**
Allow other devices (e.g. LK) to make use of MPIManager as well.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-87683]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>