Remove flaky tests and add local test seeding mechanism (#1172)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Indeterministic tests (probabilistic tests from shots), even with using
flaky test infrastructure, can some time trigger failures in CI, and
requires re-running, which increases testing time.

**Description of the Change:**
Similar to work in
[Pennylane](https://github.com/PennyLaneAI/pennylane/pull/6435) , we add
the ability to control the seed for tests. The seed can be updated
regularly to make sure we're not using lucky seeds. For relevant tests,
we pass the seed to both the default and/or lightning devices to ensure
we get deterministic results.

We also remove flaky tests. 

**Benefits:**
Reduced testing time and reduce CI failures.

**Possible Drawbacks:**
Need to regularly update rng_salt.

**Related GitHub Issues:**

[sc-85986]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>