Update stable docker build CI (#1188)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Since Catalyst updated the Device API, v0.41 does not work with
up-to-date Catalyst. v0.41.1 has the correct pinning to stable Catalyst
v0.11, but for our docker build CI test we were testing with v0.41.0
(which fails to build because of the incompatible Catalyst), instead of
v0.41.1.

**Description of the Change:**
Use v0.41.1 for Pennylane and Lightning stable docker CI.

**Benefits:**
Stable docker CI test passes again:
https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/15759901310

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-93758]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>