Distributed Lightning-Kokkos simulation with MPI (#1114)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
This PR adds support for MPI-distributed statevector simulator with the
Lightning-Kokkos simulator.

**Description of the Change:**
Supported features include:
- Forward circuit simulation
- Adjoint differentiation
- Expectation value / Variance / Probs 
- Shots: Samples/Counts

Not yet supported:
- native Mid-circuit measurements
- Sparse Hamiltonian

**Benefits:**
Ability to run large scale simulation on HPC systems with e.g. AMD GPUs.

**Possible Drawbacks:**

**Related GitHub Issues:**

**Notes for reviewers:**
you can run the tests locally with MPI on CPU.

To test locally for CPP tests: `PL_BACKEND=lightning_kokkos
build_options="-DENABLE_MPI=ON" make test-cpp-mpi`.
- kokkos specific tests need to be run with `mpirun -np 4` , and general
tests need to be run with `mpirun -np 2` (same as LGPU-MPI). This is
already set in the makefile.

To test locally for Python tests: 
- Similar to LGPU-MPI
- ` PL_BACKEND=lightning_kokkos build_options="-DENABLE_MPI=ON" make
python`
- `mpirun -np 2 python -m pytest mpitests/`





[sc-92661]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Joseph Lee <joseph_lee@login07.frontier.olcf.ornl.gov>
Co-authored-by: ANTH0NY <39093564+AntonNI8@users.noreply.github.com>