Test `set_shots` compatibility (#1173)
**Context:**
Along with the `set_shots` transform, `PennyLane` is going to decouple
the `shots` information from `device` to the application of transform
very soon. Before deprecating the `device.shots` we need to test if any
other packages in our ecosys suffer from hidden issues/bugs as well.
With enough number of tests directly using `set_shots` to configure
`_shots` information on `QNode` and passing Lightning CI, we are
convinced that Lightning is compatible with PennyLane adjustment and
ready for a further deprecation on `device(..., shots=...)` interface in
next quarter.

**Description of the Change:**
Tests adjustments

**Benefits:**
Partially done some work for future.

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-92563]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>