Update AmplitudeEmbedding tests for LK and LGPU (#1192)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Previously LK and LGPU does not support AmplitudeEmbedding after
operations. Since support has been added, the test should no longer
xfail.

**Description of the Change:**
Remove xfail for LK and LGPU for AmplitudeEmbedding test.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-76161]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>