Fix Lightning Kokkos CUDA test failure with RelWithDebugInfo due to Kokkos Lambda (#1194)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Previously for Lightning-Kokkos, when compiling for CUDA with
RelWithDebInfo , a segmentation fault occurs during tests due to using
`parallel_for(..., KOKKOS_LAMBDA)` directly in `applyMultiRZ` routine.

**Description of the Change:**
- Use Kokkos functor class instead of kokkos lambda
- Simplify some code with `using KokkosComplexVector =
Kokkos::View<Kokkos::complex<PrecisionT> *>;`
- Update LK-CUDA and LK-MPI-CUDA CIs to use RelWithDebugInfo

**Benefits:**
- No more segmentation faults for compiling LK with CUDA in
RelWithDebugInfo

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-80646]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>