Skip applying Identity for Lightning-Qubit (#1212)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Lightning qubit only accepts identity with 1 wires, which is not
necessarily the case from Pennylane - especially from Catalyst.

**Description of the Change:**
Remove assert for wires not equal to 1 for Lightning Qubit applying
identity gate.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
https://github.com/PennyLaneAI/catalyst/issues/1175

[sc-95544]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>