Add "device" as a supported MCM method that resolves to "tree-traversal" (#1210)
**Context:**
It was previously determined that `mcm_method="device"` is the
convention for using the device-native mcm method, one that does not
require any transforms on the PL side. Now that tree-traversal is
supported on the lightning devices, it can be considered a device native
MCM method.

**Description of the Change:**
Update the Lightning devices that support tree-traversal to treat
`mcm_method="device"` as tree-traversal.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-95418]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Luis Alfredo Nuñez Meneses <alfredo.nunez@xanadu.ai>