Remove squeezing from sample measurements (#1226)
Note that we expect the device test suite to fail until pennylane master
has been updated.

**Context:**

In pennylane PR
[(#7944)](https://github.com/PennyLaneAI/pennylane/pull/7944), I am
removing the squeezing of singleton dimensions from sample results. This
greatly simplifies the handling and postprocessing of sample results.

In that PR, I xfail a couple of tests that were failing because
lightning had not yet been updated. Once we merge that PR, we can merge
this one, and then stop xfailing the relevant PL tests.

**Description of the Change:**

Adjust the lightning device to this change.

**Benefits:**

PL-core behavior for samples will match catalyst, and the postprocessing
of samples will be much easier to manage.

**Possible Drawbacks:**

This is a breaking change, even if it is one we really need to make.

**Related GitHub Issues:**

[sc-97445

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Andrija Paurevic <46359773+andrijapau@users.noreply.github.com>