On-board lightning to using blacksmith runners (#1241)
**Context:**
Change the large runners that pennylane-lightning uses.

This would swap any instance of GitHub based large runners to the same
large runners offered by a third-part vendor
[blacksmith](https://blacksmith.sh).

**Description of the Change:**
This pull request changes the name of the runner returned when a large
runner is requested. In the case of lightning, this is most workflows
that use linux runners.

**Benefits:**

- Lower costs, blacksmith offers identical environment setup to GitHub
large runners but at half the price per minute. So even without
performance benefits, we slash monthly costs for CI in half.
- Faster vcpu clock cycle and higher single thread performance. From my
coarse testing, the entire test suite has had a 30% speed-up. Tests that
were running in the 18-20 min range ran in 10-12 min. Tests that ran in
12-15min ran in 6-9 min.
- Faster runner boot up. A substantial speed up for urgent PRs waiting
for a runner.
- They also offer a dashboard for all jobs run on their runners. More
details regarding this coming soon.

**Possible Drawbacks:**

The new large runners should be a complete drop-in replacement for the
existing GitHub large runner.

On the pennylane side, one test had to be updated. See
PennyLaneAI/pennylane#8134 for details.

The issue on the pennylane side did seem to be instability in the test
itself. It failed on higher versions of python across all runner
variants. Reducing the tolerance did fix it across the board.


**Related GitHub Issues:**
None.

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>