Remove `--no-deps` for PL installation for LQ stable test (#1245)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Stable/Stable CI tests should use the dependencies from stable version
of Pennylane. Currently for LQ tests, when using stable, we install
stable Pennylane with `--no-deps` which does not reflect the old
requirements. When PL updates their requirements, this breaks
stable/stable tests, e.g. autoray in this case.

**Description of the Change:**
Remove `--no-deps` when installing PL for LQ stable tests. This is
already the case for other device tests.

**Benefits:**
Restore balance to stable/stable CI.

**Possible Drawbacks:**

**Related GitHub Issues:**


[Stable/Stable](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/17328371451)
[Latest/Latest
](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/17328373930)(LG
& LT fail expected)

[sc-98480]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>