Remove randomness for LQ MCMC (#1252)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Earlier we tried to enable fixing the seed for lightning devices, which
enables deterministic measurement results. However, we have still
observed some stochastic failure due to indeterminism coming from LQ
MCMC. This is because MCMC's implementation creates separate RNGs for
the transition kernels, which is not fixed by the seed.

**Description of the Change:**
Use a reference of the same RNG device for the transition kernel to
ensure the deterministic state.

**Benefits:**
Remove indeterminism coming from MCMC tests in CI, reduce
stochastic/flaky failures
([e.g.](https://github.com/PennyLaneAI/pennylane-lightning/actions/runs/17437540026/job/49511648504#step:16:186))

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-98597]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>